Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Beto, thank you for taking care of this. I'd think that delay 60s is too small because we see in the logs "failed to drain node : worker-2 after 1 hour". Probably, a better strategy would be to check mcp status as you proposed.
> Something like this (I've tested regex on the logs output you've provided)
> ```
> - name: Register MCP status      
>   shell: >        
>     {{ oc_tool_path }} get mcp
>   register: output
> 
> - name: Check if master or worker are still updating
>   set_fact:
>     still_updating_worker: "{{ output.stdout | regex_search('worker[\\s]+rendered-worker-[a-z0-9\\s]+False[\\s]+True') != None }}"
>     still_updating_master: "{{ output.stdout | regex_search('master[\\s]+rendered-master-[a-z0-9\\s]+False[\\s]+True') != None }}"
> 
> - name: Fail if the update is still ongoing
>   fail:
>     msg: "Master or worker are still updating"
>   when: still_updating_master or still_updating_worker
> ```

The delay of 60s is to give time the mc to render the new config. It may read as too short but in most cases (we do the same in example-cnf) it's enough time. Once a new mc is rendered it's applied to the mcp and the nodes start the process of upgrade to the new mcp.  Let's test with this delay and see if it would be enough to avoid the issues seen later in the job.

Patch-set: 1