Create patch set 2
Uploaded patch set 2.

Patch-set: 2
Subject: Add validation for MCP updates  - Adding a new validation method to check that the MCPs have been updated  - The tasks to "Wait for ALL nodes to be in a Done state" may not bee needed by won't harm to keep it for some time  - Found that the MCs are created as part of the performance add on change, not during the sriov profile change  - Adding the validation to the upgrade, here the MC of to update imagecontent source policies that may not require a reboot or may not have changes.
Commit: a416acf270244910fc60c0332d9152e5a94ba5a6
Tag: autogenerated:gerrit:newPatchSet
Groups: a416acf270244910fc60c0332d9152e5a94ba5a6