Standarize tasks to use dci_pullsecret_file var
Currently there are tasks that use dci_pullsecret_file
some others use the path of the clusterX-pull-secret.txt
and some combine registry-auths and pullsecret, so this
change get creds from a single file, create it if it
does not exists, and all tasks that require it use it.
Main README has been updated to reflect these settings.

Change-Id: I83d7e2373c743f22930106142296b8a3cb576629