Also check MCP before installing performance profile
Currently, all the daily jobs where OCP version is higher than OCP 4.8 (at least for the moment) are failing when checking the MCP status after applying the performance profile with multiple hugepages.

Right now, we are assuming the original MC created for the workers is installed and running in the MCP before installing the performance profile, but it may not be.

With this, we will see if, by doing the same MCP check before installing the PerformanceProfile, we fix this issue, confirming it's a synchronization issue.

Change-Id: Ia35a0450dff0f90031ad40c2d1ab6c67debc029d