Revert performance-profile for OCP >= 4.8
Currently, all the daily jobs where OCP version is higher than OCP 4.8 are failing when checking the MCP status after applying the performance profile with multiple hugepages.

As this performance profile is currently being used just in OCP 4.7, this change reverts that multiple hugepages config as a workaround to make OCP >= 4.8 works, while checking a proper way to define the performance profile with multiple hugepages.

Change-Id: Ia35a0450dff0f90031ad40c2d1ab6c67debc029d