Generalize usage of ocp_on_libvirt
Currently, we are only admitting that ocp_on_libvirt is executed in localhost, but it may happen, in different lab configuration, that we want to run ocp_on_libvirt over a different server.

For enabling this, we need to change the hosts in which libvirt_up and libvirt_destroy are executed. By putting "all", all the servers included in ./inventory/hosts will run the tasks

Change-Id: I0538c19ce34378f9be6cc62fb5c69689bc470346