adding a variable to control catalog source update
The extra variable will permit to specify if the
CatalogSource gets updated. True by default, but
using update_catalog_channel=false it will be disabled.
This comes into play if we upgrade OCP from 4.a to 4.d
and we do not want to upgrade the operators in 4.b and 4.c,
only when getting to the latest desired OCP version.

Change-Id: I8cf649415ea126ac0cf27c8b93041394d21d0b2a