wait for perfprofile generates rendered mc
This patch removes the pause of 60 seconds to wait
for the MCP to be triggered, because sometimes is
takes longer, and checks for the Kubelet Config
timestamp to compare agains the latest rendered MC,
then it continues the normal check in wait-mcp role

It also orders some tasks that can be used for both
SNO and standard OCP clusters when applying the
performance profile. And renames some others.

Change-Id: I78e6164d32d65539801e77d2224fa31425151338
Test-Hints: sno