Add option to set installplan approval mode
Allow to set the approval mode in the operator
subscription by defaul is not required. Ommiting
this will set the Automatic mode, otherwise should
be set to "Manual".
For PAO, the variable pao_installplan_approval_mode
will pass the mode to the approval_mode var in the
OLM role which installs any operator.

Test-Hints: sno
Change-Id: Icc5cdddd15fdc83fa66d11fd19a5e0e5da90fd5a