move tasks to role for tracking rpms and git repos
This patch integrates the independent plays and
tasks in a role in the common directory so it can
be reused by dci app agent too, and have a common
place to pull the software information and add it
as components. Minor modifications were made for
the role to use its own variables and get them
in the pre-run play.

Test-Hints: sno
Change-Id: I582fc8c2b7c06858c4edb86cf27e736da3d10037