Monitoring operators upgrades
Once a new catalog is available, the operators are automatically upgraded if the subscription is set to Automatic.
We need to monitor the process for each operator to reach the new CSV defined in the new bundles.
Some operators may require multiple jump to reach the latest version. For example CNV.

* Setting the CSV validation task to ignore errors to test how it behaves with multiple operators during the testing.

Change-Id: Ie3ce20709c321c660bac5da7a7fc3bcbcf4305b2