Consolidate wait-mcp and check-sriov in one single common role
Currently, we have one role to check MCP and we use another role in some partners' repos to check SRIOV resources, but both of them are mostly following the same structure in terms of code.

This change homogenize both roles to merge them in one single role, called check-resource, that is executed a specific amount of times if we need to apply the workaround, as it could happen that, even applying the workaround, some nodes move back to SchedulingDisabled status after a while.

Change-Id: Ic436678302265c43c4c8a4518f8fde6433dcd950