Updates to test-runner to support assisted in new lab
Update to test-runner to use the prefix logic which will use the correct
config in the new bos2 lab.  Re-using the same logic from the sno
samples to generate the hosts file.  We may be able to merge some of
this in the future once we get this working first.

Test-Hints: assisted
Test-Args-Hints: -e "dci_components_by_query=['tags:build:ga']"
Change-Id: Ia80dc57b82b3fd0796b9fcf7685a0045690a1ba9