Integration for multiple operators.
When installing multiple operators under the same namespace, if any of
the subscriptions is installed with install plan approval set to
'Manual', it forces every subsequent subscription to get the install
plan approval set to 'Manual'.

The problem is the last task of the olm-operator role patches the
install plan and sets the approval to 'Manual' by default.

This breaks our workflow since, for the next operator installed in the
same namespace, the job will block until the CSV is created, which won't
happen until the install plan is approved.

The idea behind this CR is to have all subscriptions created with
install plan approval set to 'Automatic', even if it'll be moved to
'Manual' afterwards, when the install plan is patched.

This CR also:

- Documents the usage of variable "ns_labels".
- Parameterizes the operator group name with variable
  "operator_group_name", yet defaulting to "operator".
- Provided and example of using Test-Args-Hints in development.md

Test-Args-Hints: -e dci_prepare_cnf=True
Depends-On: https://github.com/dci-labs/keysight-lab-config/pull/28
Change-Id: Idc9593f4d3d09ace1672050e9c710717ec91f4a0