Integration for multiple operators.
In the initial solution, a subscription is created, then the job waits
for CSV to be created and finally the install approval is patch so it may
be changed from "Manual" to "Automatic".
The problem is the CSV won't be created until the Install Plan is
approved which, when set to "Manual", won't happen. Thus the need to
change the order of the tasks.
For some reason, this issue is only observed when installing more than
one operator in the same namespace, since the first Install Plan is set
to "Automatic" by default.
This CR also:

- Documents the usage of variable "ns_labels".
- Parameterizes the operator group name with variable
  "operator_group_name", yet defaulting to "operator".

Change-Id: Idc9593f4d3d09ace1672050e9c710717ec91f4a0