Patching subscription channel in upgrade-csv.yml
During the upgrade process, the operators are upgraded as well.

In its current version, the code assess the status of the operator's
subscription, but it only focuses on making sure the install approval
plan is 'Automatic'.

However, during the process, the job infers the new version CSV and
blocks until it's created. But for the subscription to trigger a new CSV
creation, the channel must be set to the new version.

This issue is currently not a blocker because the task is set to ignore
errors. However the loop takes 1 hour per operator to timeout.

This changes add this channel update.

- After Beto's advice, moved from using ocp_version to default_channel.

Test-Upgrade-Hints: yes
Test-Args-Hints: -e enable_perf_addon=true -e enable-acm=true
Test-Upgrade-From-Topic-Hints: OCP-4.9
Test-Upgrade-To-Topic-Hints: OCP-4.10
Test-Upgrade-Args-Hints: -e dci_name=csv_upgrade_4.10
Change-Id: I52757cad443046d3a15113f662e643666d90d108