Update dci-check-change to use existing inventory by default
Currently dci-check-change will overwrite the system inventory file
which is not what we want.  We should be able to tell a partner to run
dci-check-change with a pr and it should run with their current config.

WIP - This is just echoing the current sudo cp to try and understand
what is happening.

Change-Id: I50d629d402183c7efae58c50c3a73435570be117