Decouple performance profile definition and operators installation for CNFs
In OCP 4.11 and 4.12, we are not installing performance-addon-operator but we are installing a performance profile, and it is done because enable_sriov is set to true.

However, this dependency should not be present; we may want to install a performance profile even though we are not using SRIOV.

This change takes care of decoupling these two elements.

Change-Id: I241bb351104db0217ebbc9070b1d4240be474ced