Adding storage-tester role to test storage service during the upgrade
It is launched only during an upgrade if the variable storage_upgrade_tester is set to true.
A default Storage class must be set otherwise the volumes creation will fail.

At the end of the upgrade, info are gathered from the storage-tester namespace.
Then, logs are pushed in DCI as 3 log files and a junit report.

See the documentation for more details

Change-Id: Icacd97939b6f90c498db6e8f8e571c53054945da