Checking new CSV versions from same catalog source
It may happen that same operator is provided by 2 different catalog
sources.

This would cause a conflict if the next CSV version is retrieved from a
catalog source different than the installed operator.

This change allow for operators to be available in different catalog
sources and guarantees the next CSV version is retrieved from the same
catalog as the installed operator.

- Applied Beto's suggestion on using label selectors.

Change-Id: Ib3b9cc12fc90fde59650414f5eda894b5e4d106c