dci-check-change/test-runner - parse kubeconfig independently
If kubeconfig and prefixes are used when running dci-check-change, it is not working because, according to dci-check-change docs, kubeconfig has to be included prior to any other arguments (like prefixes with -p and -p2). But then, on test-runner, it expects to receive the kubeconfig after prefixes, so it's not working.

This change separates the processing of kubeconfig and includes it at a specific position for test-runner, so that we don't have this issue.

Change-Id: If17b4527143c75bff8d0b46b46dc1c2758c80a64