dci-check-change/test-runner - parse kubeconfig independently
If kubeconfig and prefixes are used when running dci-check-change, it is not working because, according to dci-check-change docs, kubeconfig has to be included prior to any other arguments (like prefixes with -p and -p2). But then, on test-runner, it expects to receive the kubeconfig after prefixes, so it's not working.

This change allows to process kubeconfig at the correct place in test-runner, so that we can parse it together with prefixes

Other changes that were needed to make all the components to work fine:

- Updating docs to reflect the usage of prefixes and kubeconfig altogether
- Changing PREFIX_NAME and PREFIX2_NAME value, as they are not currently being used, and using PREFIX2_NAME afterwards
- Update CONFIG_REPO when moving to dci-openshift-app-agent config to do the same than with dci-openshift-agent, else it will not work properly
- If prefix for dci-openshift-app-agent is used, take configuration variables from the settings file related to the prefix
- Remove -d option when running dci-openshift-app-agent-ctl in one of the executions, as it is not used in other commands and it was not working

Change-Id: If17b4527143c75bff8d0b46b46dc1c2758c80a64