run after-install in its own step to avoid delegate
remove delegate expression that make file fetching
complicated. This play is executed from the localhost
now for IPI, UPI and SNO, it should fetch the files from
the provisioner node and place them in the ansible
controller. AI will have a similar task but executed
from the jumphost, so the fetch is more a copy from
directories in the localhost.

Change-Id: I21bc38c5a5e28816fb2baee35332c29ae57609cb