Dynamically use mirror mapping for mirrored images
The imageContentSourcePolicy file is not used nor generated
from the mirroring process.
Currently the imageContentSources are coming from a file
(disconnected_registry_mirrors_file), with this change the content will
come from the mirroring process and used in the cluster.
Always mirror the images, this allows to generate the right
imageContentSources every time, even in an upgrade.

Change-Id: Ib2ac09364ae6edd2340b4869f289455273842c5c