Add version to openshift-client-linux.tar.gz in SNO and UPI
This was missed when updating the agent for allowing nightly builds. SNO on libvirt is not working because of this.

For assisted, Crucible project needs to be changed, this will be done in a separate change on that repo, testing it in https://softwarefactory-project.io/r/c/dci-openshift-agent/+/27118

Test-Hints: sno

Change-Id: I82e13b395b683bd99f2b42721f0c1fd964fd2c6f