Add support to connected nightly builds
The baremetal-deploy in a connected environment pulls artifacts from a
release_url. This change removes the need to use the url for nightly
builds as is not reliable. We provide a webserver from the controller
using the already obtained artifacts and provide it to the provisioner
to consume.

Test-Hint-Args: -e dci_topic=OCP-4.11 -e dci_components_by_query:['type:ocp,tags:build:nightly']
Change-Id: I2c5ad4fbf5eb16d27ccc92056949c53e159f01ec