Identify right pkg manifest when upgrade operators
When installing operators from custom catalogs, the same package
manifest may be available in different catalogs.

When the package manifest was retrieved in order to define the operator
upgrade path, only one instance was returned, even if there were
serveral matching manifests so, if the wrong manifest is returned it may
not be possible to create a valid upgrade path and the job will fail.

This change retrieves the extra field catalogSource from the existing
subscriptions and then looks for the package manifest matching both the
operatorName and the catalogSource.

- Fixed hardcode namespace name as spotted by tgarcia.
- Fixed wrong variable definition: was "catalog" should have been
  "catalogSource".
- Filtering by field name when getting packagemanifest_all.
- In role olm-operator getting the currentCSV id from the subscription
  instead of the package manifest to avoid problem when working with
  multiple manifests for the same operator.
- As Beto suggests, using the subscription as the source of trust when
  identifying the current CSV.

Depends-On: https://github.com/dci-labs/intel-nm-lab-config/pull/144
Change-Id: I804fba6563a84d0ef37472abe475a836078b9988