Installing specific operator versions
This change adds the variable starting_csv to the olm-operator role.

The variable must be set to a valid operator version from the current
package manifest.

When the variable is defined and the operator is installed, the agent
will add the startingCSV parameter to the subscription and force the
approval method to be manual, to prevent any automatic updates to the
latest version.

Then the role will try to identify the install plan for the operator and
approve it, so we make sure only the target version install plan is
approved.

Test-Args-Hints: -e {"dci_operators":[{"name":"mtc-operator","catalog_source":"redhat-operators","namespace":"openshift-migration","starting_csv":"mtc-operator.v1.7.5","ns_labels":{"openshift.io/cluster-monitoring":"true"},"operator_group_spec":{"targetNamespaces":["openshift-migration"]}}]}

Change-Id: If5a22e3cdf37bba3adb83b670ead32db91b9b9b6