Installing specific operator versions
This change adds the variable starting_csv to the olm-operator role.

The variable must be set to a valid operator version from the current
package manifest.

When the variable is defined and the operator is installed, the agent
will add the startingCSV parameter to the subscription and force the
approval method to be manual, to prevent any automatic updates to the
latest version.

Then the role will try to identify the install plan for the operator and
approve it, so we make sure only the target version install plan is
approved.

Test-Args-Hints: -e
'{"custom_catalogs":["quay.io/telcoci/nfv-example-cnf-catalog:v0.2.9"]}' -e
'{"dci_operators":{"name":"testpmd-operator","catalog_source":"nfv-example-cnf-catalog","namespace":"example-cnf","operator_group_spec":{"targetNamespaces":"example-cnf"}}}'
Change-Id: I289542745051f0462ef0b97d08c8977b75a499ca

Change-Id: I8dc7b1b7a4cf7d408aa177a517864cbba9d4357a