Resuming validation of CR 27314
CR 27314 was approved and merged, but it introduced a bug in the form of
a quoted omit variable "default('omit')" in the olm-operator labels
definition which prevented the operators not defining any label to fail to
install.

We reverted the CR and resumed its validation here.

Test-Args-Hints: -e
{"dci_operators":[{"name":"sriov-fec","catalog_source":"certified-operators","namespace":"vran-acceleration-operators","operator_group_spec":{"target_namespaces":["vran-acceleration-operators"]}}]}
Test-Upgrade-Hints: yes
Test-Upgrade-From-Topic-Hints: OCP-4.9
Test-Upgrade-To-Topic-Hints: OCP-4.10

Change-Id: I22463b4b81a5d8956eef3b28726e143f82d0776a