[DNM] Testing real time kernel upgrades
This is just a testing support CR needed to see if the latest versions
of OCP support upgrading clusters with the worker nodes running on top
of the realtime kernel.

We created this CR because the testing depends of the changes from CR
27802 which is ready to be merged and we don't wont to add any noise to it.

We'll be testing this in the NM lab.

Change-Id: I36878d18c99033b14f379c20033465439d96547e