Include workaround for OCPBUGS-11779, refactor workaround for OCPBUGS-6676
We have seen workaround for OCPBUGS-6676 is useless as there is at least one node in NotReady status, so we cannot do anything with that. So, in case it happens, just fail at that point.
For OCPBUGS-11779, as we are covering again the case of Ready,SchedulingDisabled status, then apply the uncordon to that node

Change-Id: I7396be349f49e4b9f00d51077baf345dd58818d6