Assisted - increase final ClusterVersion check retries
There are some sno-ai tests where this task was failing because ClusterVersion was not in Completed state, however, when checking the logs, it looks like it transitioned to Completed while logs were retrieved, which takes around a minute. So, increasing the number of retries so that we have one extra minute to check.

Test-Hints: sno-ai

Change-Id: I44dda425264144f8355ff36f11e023a82aa48255