Use ImageDigestMirrorSet for 4.14+ in disconnected
imageContentSources is being deprecated in openshift-installer
imageDigestSources is taking its place in a disconnected environment
when mirrors need to be set to the cluster

ImageContentSourcePolicy (icsp) resource is being deprecated in 4.14
ImageDigestMirrorSet and ImageTagMirrorSet are taking its place, for now
we are only adding ImageDigestMirrorSet (idms) support.

Adds support of ImageDigestMirrorSet when using a catalog from file,
this play is using oc-mirror that is producing an icsp, thus adding the
support for that case.

In this change we are adding support to idms in 4.14, some integration
to idms is still in development at this time.

Change-Id: I68ac896b3a4622fb719a46610e881c54b61d66aa