extract-dependencies | Improve the check of PRs already merged
Current code is not detecting if a merged PR is, effectively, merged. This will use the JSON file retrie
ved from Github and check the merged attribute to see if the change is merged or not. In case it is merg
ed, then extract the merge_commit_sha and checkout to that commit to really test what it has been merged
.

Depends-On: I2f360bbb5ff62f05a3535375e7c07081cd3feb19

Change-Id: I3410b86d64a69117a6cedcc41e679f50a5119a3e