extract-dependencies | Use Github-Gerrit change info to rely on its merge status
Current code is not detecting if a merged change is, effectively, merged. This will use the JSON file retrieved from Github-Gerrit and check some attributes to see if the change is merged or not.
In case it is merged, then checkout to master branch to test the latest. We don't checkout to the merged commit just in case we leave some depends-on in some changes, just to always test the latest.

Depends-On: I2f360bbb5ff62f05a3535375e7c07081cd3feb19

Change-Id: I3410b86d64a69117a6cedcc41e679f50a5119a3e