test-runner | run libvirt_destroy from ocp_on_libvirt prior to AI jobs
If an IPI installation fails prior to an AI job and the bootstrap VM from IPI has the API VIP still there, it may affect to AI deployments.
As we're using the same naming convention for all installers, let's clean the VMs also prior to starting with AI, and this will include the provisionhost that is not deleted in current AI logic.

Change-Id: I883881196bac9f4f6bb41c30c00010db7443d84b