Adding sushy tools to the ocp-on-libvirt solution
Currently ocp-on-libvirt installs the conserver service along with the
VMs to allow for IPMI remote management of the hosts, which in turn
allows for PXE image boot up.

In order to support ZTP deployments, virtual media image boot up is
needed, and to manage it remotely, the target server BMCs must suport
the redfish protocol.

Sushy tools provides support for the redfish protocol for libvirt VMs.

This change installs sushy tools, as implemented in the
assisted-installer ansible project,  when running the ocp-on-libvirt
role.

- Added dci-vault to libvirt_up call
- Enabled virtualmedia in ocp-on-libvirt

Depends-On: https://github.com/dci-labs/bos2-ci-config/pull/72
Depends-On: https://github.com/openshift-kni/baremetal-deploy/pull/975
Depends-On: https://github.com/redhat-partner-solutions/crucible/pull/272
Test-Args-Hints: -e enable_redfish=true -e enable_virtualmedia=true
Change-Id: Ie862c539baaee9872d3a3c1ae06a696f9723550d