Allow passing proxy variables in inventories
Not only allow proxy env variables but allow inventory variables to be
used to define proxy vars, otherwise group_vars/all takes over the
inventory ones.

This is meaningful for IPI, UPI and SNO installs that use a provisioner.

Also modify dnf in provisioner when an http_proxy is defined a
disconnected provisioner is unable to install packages because the
http_proxy environment is not used in the package ansible module,
enabling it globally in the provisioning will help to make use of it in
following tasks

Add a variable to enable/disable proxy easily to allow define
http_proxy, https_proxy and no_proxy_list in the inventory without
activating them unless explicitly enabled.

Renaming variables and documenting them in the main README.

Change-Id: I997337d914d31c23a0ce97defa38afb5a90600d2