Do not fail if the a the performance profile already exists
The task get the timestamp fails if the performance profile already exists because the is no change. Removing this condition will allow to get the timestamp of the already created profile.
The other tasks will be skiped because there is no actual change in the MCP and the creation time is < than the defined thresholds

Change-Id: I2aead2b8b4e4f3f9f4f0ecdf23529cd5418c7505