Relabelling artifacts container mounted files
When deploying an OCP nightly build, an ad-hoc web server container is
run to server the release artifacts from the jumpbox local filesystem.

Labelling may prevent the artifacts container to get authorization to
access the release artifacts it's set to serve.

This change forces the proper relabelling of the mounted files so the
container will be able to read them.

Test-Args-Hints: -e dci_topic=OCP-4.15
Change-Id: Ia72e674d9b226da89230de53103c5c358dbb9dbb