Remove wrapper script opm-auths
The opm-auths is a wrapper script that brings support for multi-entry
registry authentication when using opm. The opm client lacked the
functionality as reported in:
https://github.com/operator-framework/operator-registry/issues/935

But lately it was added in:
https://github.com/operator-framework/operator-registry/pull/1165
And now it is available in the stable ocp clients
https://mirror.openshift.com/pub/openshift-v4/x86_64/clients/ocp/stable/

Thus, the wrapper is not needed any longer.

Lastly, opm stable now provides two FIPS compiant binaries. ONe for rhel8
and another for rhel9, downloading the correct one for the agent.

Change-Id: Idab1a56a594a2f5c1e6a623af8d6d9b7ff359cb1