Trimming space characters in APP_NAME
dci-check-change is misbehaving when testing PRs from github and the
Test-App-Hints setting is provided.

This parameter is parsed including a trailing vertical space character
which prevents the tool from properly defining the app agent jobs.

With this change we trim any vertical and horizontal space characters
from the app name passed through the Test-App-Hints parameter.

Change-Id: Icaa317ba7b3e9fb5d8bb62aa77cbcf1ad3bbb510