Provision host does not need EPEL release
Currently, when configuring the provision host, the agent checks and
sets the EPEL release.

This is OK for hosts not based on RHEL, but with RHEL this task is
likely to fail.

To avoid this, we add an extracondition to the task so EPEL is only
checked when the OS distribution is different to RedHat.

Change-Id: I385956bc62a42a2d56fdbfa070d10c4d7b4694b4