Added new deployment method acm - ztp-spoke
If there's an ACM hub cluster available, this change uses the hub
kubeconfig file (hub_kubeconfig_path variable) to configure the cluster
in order to connect it to a  git repository containing the site config
manifest needed to deploy a spoke cluster. it then monitors and waits
for the spoke cluster deployment to conclude.

Change-Id: I5364e0cae2557a7846644b348c6637ce7a1c7e45