acm-ztp-spoke | Wait a maximum of 15 minutes for the AgentClusterInstall to be created
There are cases like this one [1] where it is not enough to wait 5 minutes for the AgentClusterInstall to be created. In fact, the cluster was created, but DCI did not detect that this resource was created "on time".

Waiting some more time to be able to check the AgentClusterInstall creation. Less than 10 minutes should be fine but setting 15 minutes just in case.

TestBos2: bevo/acm-hub bevo/ztp-spoke

Change-Id: I8a3a84358f02ea1ffd019d9e885df09843d88d64