Add flag to enable a catalog with upstream operators
In the past we build upstream operators (SR-IOV) for non-GA builds as
the index/catalog took really long time to be created. But now that we
generate Pre-GA operator catalog/index we have early access to operators
in pre-ga builds.

Making the use of a catalog with upstream operators optionals through a
flag rather than depend on an OCP version.

Change-Id: I09fd55f7bece2fab448c5a12c5a6e60ca494f83b