Add support for performance profile during install
The support of performance profile during install was added in OCP 4.13,
but we are adding it for 4.14+ and moving tuned profile support to its
own play.
Increasing the healthcheck after install as with PAO it will take a bit
longer to apply it to all the nodes.

TestDallas: ocp-4.15-vanilla
Change-Id: Ie86fad92925c1b0791a13bef1d7bbc503ceb253d