Add a way to tag pre-ga images
Allow to tag pre-ga images that succesfully completed a job.
It requires that the job uses a pre-ga component and an authfile
provided to push the tagged image back to the registry.

The tag adds a -dci to its original tag to reflect that it was tested
succesfully by DCI.

Change-Id: I81cbf4b39c00bdfa8dace9f33f8af7e863e32d2c