Enabled KSOPS feature for GitOps
To run ZTP/GitOps based deployments of a spoke cluster, the pull-secret
and baremetal host BMC credentials must be provided to the hub cluster.

This data is considered sensitive and must not be shared openly over Git
repositories.

With the KSOPS plugin, the GitOps operator may decrypt data stored
encrypted in the repositories it synchronizes with, thus allowing us to
keep the process fully automated.

Depends-On: https://github.com/redhatci/ansible-collection-redhatci-ocp/pull/648

Change-Id: I824c8ec0fde9323d78c127ee4869773c64f8fe14