Increase time for ZTP deployments / set kubeconf in jumphost
For SNO it has been observed that it may take about 60min to deploy the
cluster, making it too close to fail. Increasing 10 more min.

Setting hub_kubeconfig_path in jumphost

The hub_kubeconfig_path is defined from a previous job through dci-pipeline
in its inputs/outputs. But when a job is running on top of a deployed
cluster it uses HUB_KUBECONFIG env var, this is a requirement for ACM
install types (SNO, HCP, ZTP).

In the case of hybrid clusters (kvirt) it is not needed a
hub_kubeconfig_path, removing that code.
We don't define a KUBECONFIG in the pre-run with hub_kubeconfig_path
because that is only used in the ACM pre-run, but we include KUBECONFIG
only if it is in the environment.

Change-Id: I2835c0695b77828bc10ff5144f06d9ac8ff62976