Use acm.utils role to generate registries.conf
The registries.conf is required to install disconnected clusters through
ACM when the version installed is diferent to the one the Hub is
running.

This is only implemented in SNO, but by moving to a role it allow us to
use it in other ways to install ACM like clusterInstance.

Change-Id: I02bc498983f499a494c6f0d0e5ef4a773a578e15