Installing additional CA certificates for Image Service by default
If the additional trust bundle is added to the install-config, it's safe to asume it may be added to the Image Service containers, regardless of the cluster being disconnected.

This change is required by an environment placed behind an internal security proxy interferring in the TLS certificate exchange with the OpenShift public mirrors.

For the downloads to work, the involved systems need to accept the internal proxy certificates. This means the environment is not disconnected, yet the certificates must be installed.

TestBos2Sno: sno
Change-Id: I629d3dd468d78eeac5d7dc45ea62d0edb4b9b112