Add an option to apply PAO after cluster install
In OCP 4.16 and above, is possible to install the performance profile
during the installation of the cluster, but in some cases it is not
ideal or not functional. In the case of compact cluster it will cause
issues. A new option has been added to allow moving the application of
the performance profile after cluster install.

Change-Id: I2b3864f3d54faeb4642496bfefddb0f1d63d9ee5