Remove policy validation
The policy validation was added for ACM deployments, but policies are
namespace bound, and the validation did not specify a namespace, making
it use the one defined in the KUBECONFIG provided, so it did not
reliably provided a validation of validations.
With support added into redhatci.ocp.acm.utils for policy validation it
is now possible to get that validation on an specified namespace.
At the moment is recommended to do it in the hook where the policy is
created instead.